+++ attributes/default.rb
@@ -12,21 +12,19 @@ include_attribute "cubrid::broker"
 # latest build numbers for each CUBRID version in the form of 'version'=>'build_number'
 build_numbers = {'9.1.0' => '0212', '9.0.0' => '0478', '8.4.3' => '1005', '8.4.1' => '7007'}
 
-# the default version of CUBRID to install
+# The default version of CUBRID to install
 default['cubrid']['version'] = "9.1.0"
-# the full version of CUBRID including the build number
+# The full version of CUBRID including the build number
 set['cubrid']['full_version'] = "#{node['cubrid']['version']}.#{build_numbers[node['cubrid']['version']]}"
-# the architecture of CUBRID binaries to install based on the current system architecture
+# The architecture of CUBRID binaries to install based on the current system architecture
 set['cubrid']['arch'] = node['kernel']['machine'] =~ /x86_64/ ? "x86_64" : "i386"
 
-# the file name of the archive to download
+# The file name of the archive to download
 set['cubrid']['filename'] = "CUBRID-#{node['cubrid']['full_version']}-linux.#{node['cubrid']['arch']}.tar.gz"
 
 # the full URL of the TAR archive to download
 set['cubrid']['tar_url'] = "https://sourceforge.net/projects/cubrid/files/CUBRID-#{node['cubrid']['version']}/Linux/#{node['cubrid']['filename']}/download"
 
-# the home directory of a Vagrant user
-default['cubrid']['user_home_dir'] = "/home/vagrant"
 # the target directory to install CUBRID
 default['cubrid']['home'] = "/opt/cubrid"
 
@@ -38,6 +36,11 @@ default['cubrid']['lang'] = 'en_US'
 # Calculated based on the recommendation from the manual page
 # http://www.cubrid.org/manual/843/en/Connection-Related%20Parameters
 default['cubrid']['max_clients'] = node['cubrid']['max_num_appl_server'] * (node['cubrid']['broker_count'] + 1) + 10
+# http://www.cubrid.org/manual/91/en/admin/config.html?highlight=memory%20parameters#memory-related-parameters
+default['cubrid']['sort_buffer_size'] = '2M'
+# default `--db-page-size` is 16KB = 16384 bytes.
+# 4 pages = 4 * 16384 = 65536 bytes = 64KB
+default['cubrid']['temp_file_memory_size_in_pages'] = 4
 
 # The configurations directory.
 set['cubrid']['conf_dir'] = "#{node['cubrid']['home']}/conf"