+++ metadata.rb
@@ -6,12 +6,14 @@
 #
 # Distributed under MIT license
 #
+name             "cubrid"
 maintainer       "Esen Sagynov"
 maintainer_email "kadishmal@gmail.com"
 license          "MIT"
 description      "Installs/Configures CUBRID Database"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "2.3.3"
+version          "2.4.0"
+
 
 depends          "build-essential"
 depends          "database"
@@ -25,6 +27,20 @@ depends          "yum"
   supports os
 end
 
+# Provided recipes.
+provides "cubrid::default"
+provides "cubrid::broker"
+provides "cubrid::demodb"
+provides "cubrid::ha"
+provides "cubrid::new_dbs"
+provides "cubrid::pdo_cubrid"
+provides "cubrid::perl_driver"
+provides "cubrid::php_driver"
+provides "cubrid::python_driver"
+provides "cubrid::shard"
+provides "cubrid::web_manager"
+
+# Recipe descriptions.
 recipe "cubrid", "Installs CUBRID Database"
 recipe "cubrid::broker", "Configures additional CUBRID Brokers"
 recipe "cubrid::demodb", "Installs CUBRID demodb database"
@@ -36,3 +52,6 @@ recipe "cubrid::php_driver", "Installs CUBRID PHP driver"
 recipe "cubrid::python_driver", "Installs CUBRID Python driver"
 recipe "cubrid::shard", "Configures CUBRID SHARD in multi VM environment"
 recipe "cubrid::web_manager", "Installs CUBRID Web Manager"
+
+# Provided resources.
+provides "cubrid_database"